backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
selenium
spacy
transformers
requests
beautifulsoup4
serpapi
python-dotenv
